Habit Modify Implemented
The "habit modify" command is now implemented.

If neither a new name nor a new periodicity has been passed, a message will be displayed and nothing happens.
Otherwise a session will be opened and the actually modified fields will be updated.
If the current values are the same as the new ones, nothing happens.

---

Validations
- validate_habit_name & validate_periodicity now allows value to be None
 - If the option is required its already checked by Click

Habit
- get() habit_id & habit_name parameters now default to None
- Added __check_streak_validity docstring